Create `Debouncer`; use it for `HatAllocator`
